C# Input with Web Application... Maybe
This is a really early build. I didn't have the current code when doing this so I tried my best to have an early start to the application that would then lead to the filtered opportunities. 

Ignore the Main.aspx.cs file... however I had to make a few changes to get past initial errors.